Took out casting in the threshold function

+++ src/org/usfirst/frc/team687/robot/Robot.java
@@ -61,9 +61,9 @@ public class Robot extends IterativeRobot {
     
     public double threshold(double input)	{
     	if(input > 1)	{
-    		return (double) 1;
+    		return 1;
     	}	else if(input < -1)	{
-    		return (double) -1;
+    		return -1;
     	}	else	{
     		return input;
     	}
